pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
